resetHelper (untested)

+++ php/helper/resetHelper.php
@@ -49,6 +49,11 @@ function resetContent()
         if(is_file($file))
             unlink($file); // delete file
     }
+
+    $file = '../../config/projects.json';
+        if(is_file($file))
+            unlink($file); // delete file
+
 }
 
 function resetSettings()
@@ -56,7 +61,7 @@ function resetSettings()
     $files = glob('../../config/*'); // get all file names
     foreach($files as $file)
     {
-        if(is_file($file) && $file != "version.txt" && $file != "colors.json")
+        if(is_file($file) && $file != "version.txt" && $file != "colors.json" && $file != "projects.json")
             unlink($file); // delete file
     }
 }
